+++ ark/dark/tsWithArkType.tmLanguage.json
@@ -48,7 +48,7 @@
 		},
 		"arkChained": {
 			"contentName": "meta.embedded.arktype.definition",
-			"begin": "([^\\)\\(\\s]+)?(\\.)(and|or|when|extends|intersect|exclude|extract|overlaps|subsumes)(\\()",
+			"begin": "([^\\)\\(\\s]+)?(\\.)(and|or|when|extends|intersect|exclude|extract|overlaps|subsumes|to|module)(\\()",
 			"beginCaptures": {
 				"2": {
 					"name": "punctuation.accessor.ts"
